Created the Solutions.md but needs Editing
I was able to add in all the features that were created this Phase. I didn't go into too much into detail about the back-end since I don't that's really required for this. I did explain all the new features add like the bar on the history page, the login and sign up and changed some of the wording a bit. Some of the paragraphs will be the same as part 2 since those features didn't really change and I doubt we need to create different wordings for the same thing. Please review and let me know if you want anything added/changed/removed.